pyproject.toml
[tool.poetry]
name = "test-pilot"
version = "0.1.0"
description = "Test scripts for Playwright MCP and Gemini integration"
authors = ["Shuhai Miao <smiao@icims.com>"]
packages = [{include = "test_pilot", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
python-dotenv = "^1.1.1"
langchain-core = "^0.3.68"
langchain-openai = "^0.3.27"
langchain-mcp-adapters = "^0.1.9"
langgraph = "^0.5.2"
mcp = "^1.10.1"
google-generativeai = "^0.8.5"
grpcio = "^1.73.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
